+++ lib/contract.mli
@@ -7,5 +7,6 @@ module Json_codec = Json_codec
 module Path_template = Path_template
 module Endpoint = Endpoint
 module Request = Request
+module Response = Response
 module Validate = Validate
 module Openapi = Openapi